move internals into a nested namespace
