+++ src/SixLabors.Core/Memory/Internals/BasicByteBuffer.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Six Labors and contributors.
 // Licensed under the Apache License, Version 2.0.
 
-namespace SixLabors.Memory
+namespace SixLabors.Memory.Internals
 {
     /// <summary>
     /// Provides an <see cref="IManagedByteBuffer"/> based on <see cref="BasicArrayBuffer{T}"/>.
