+++ src/SixLabors.Core/Memory/Internals/ManagedBufferBase.cs
@@ -4,7 +4,7 @@
 using System.Buffers;
 using System.Runtime.InteropServices;
 
-namespace SixLabors.Memory
+namespace SixLabors.Memory.Internals
 {
     /// <summary>
     /// Provides a base class for <see cref="IMemoryOwner{T}"/> implementations by implementing pinning logic for <see cref="MemoryManager{T}"/> adaption.
